Official Replica/requirements.txt
# NQ Trading Bot Requirements
# Python 3.9+ required

# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio>=3.4.3

# Data analysis
scikit-learn>=1.0.0
scipy>=1.9.0

# Logging and monitoring
colorlog>=6.7.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.18.0